Move test code out of extension
